# sass/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MIS API 文档",
        "description": "伟明环保设备有限公司SIS系统 API 文档 api/documentation  php artisan l5-swagger:generate",
        "version": "2.0.0"
    },
    "paths": {
...
        "/api/auth/refresh": {
            "get": {
                "tags": [
                    "权限认证auth"
                ],
                "summary": "刷新token",
                "description": "使用说明：刷新token",
                "operationId": "refresh",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
        "/api/alarm/page": {
            "get": {
                "tags": [
                    "报警alarm"
                ],
                "summary": "分页获取数据列表",
                "description": "使用说明：分页获取数据列表",
                "operationId": "alarm-index",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "num",
                        "in": "query",
                        "description": "每页数据量",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "页数",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "content",
                        "in": "query",
                        "description": "内容",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "succeed"
                    }
                }
            }
        },
...
        "/api/alarm-rule": {
            "get": {
                "tags": [
                    "报警规则alarm-rule"
                ],
                "summary": "分页获取数据列表",
                "description": "使用说明：分页获取数据列表",
                "operationId": "alarm-rule-index",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "num",
                        "in": "query",
                        "description": "每页数据量",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "页数",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "关键字中文名搜索",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "报警类型",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "succeed"
                    }
                }
            },
            "post": {
                "tags": [
                    "报警规则alarm-rule"
                ],
                "summary": "新增单条数据",
                "description": "使用说明：新增单条数据",
                "operationId": "alarm-rule-store",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "中文名字",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "device_id",
                        "in": "query",
                        "description": "设备ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "dcs_standard_id",
                        "in": "query",
                        "description": "标准名称ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "报警类型 communication通信, scene现场",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "enum": [
                                    "communication",
                                    "scene"
                                ]
                            },
                            "default": "communication"
                        }
                    },
                    {
                        "name": "period",
                        "in": "query",
                        "description": "周期时间（秒数）",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sustain",
                        "in": "query",
                        "description": "持续周期数",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "min_value",
                        "in": "query",
                        "description": "最小值",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "max_value",
                        "in": "query",
                        "description": "最大值",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "alarm_grade_id",
                        "in": "query",
                        "description": "报警等级",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "notify_user_ids",
                        "in": "query",
                        "description": "通知用户ID列表，英文逗号隔开",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "store succeed"
                    }
                }
            }
        },
        "/api/alarm-rule/{id}": {
            "get": {
                "tags": [
                    "报警规则alarm-rule"
                ],
                "summary": "获取详细信息",
                "description": "使用说明：获取详细信息",
                "operationId": "alarm-rule-show",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "主键",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "succeed"
                    }
                }
            },
            "put": {
                "tags": [
                    "报警规则alarm-rule"
                ],
                "summary": "修改",
                "description": "使用说明：修改单条数据",
                "operationId": "alarm-rule-update",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "主键",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "中文名字",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "device_id",
                        "in": "query",
                        "description": "设备ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "报警类型",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "period",
                        "in": "query",
                        "description": "周期时间（秒数）",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sustain",
                        "in": "query",
                        "description": "持续周期数",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "min_value",
                        "in": "query",
                        "description": "最小值",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "max_value",
                        "in": "query",
                        "description": "最大值",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "alarm_grade_id",
                        "in": "query",
                        "description": "报警等级",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "notify_user_ids",
                        "in": "query",
                        "description": "通知用户ID列表，英文逗号隔开",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "update succeed"
                    }
                }
            },
            "delete": {
                "tags": [
                    "报警规则alarm-rule"
                ],
                "summary": "删除单条数据",
                "description": "使用说明：删除单条数据",
                "operationId": "alarm-rule-destroy",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "主键",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "detroy succeed"
                    }
                }
            }
        },
...
    },
...
}